+++ src/main/java/hotel/domain/repository/RoomRepository.java
@@ -6,10 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.UUID;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
     Room findByNumber(String roomNumber);
 
     List<Room> findAllByStatus(RoomStatus status);
